+++ CHANGELOG.md
@@ -8,7 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v1.5.0] - Unreleased Phase 4
 
 ### Added
+
 - Implemented *Refresh Access Token* method with documentation and example in [README](https://github.com/signnow/SignNowNodeSDK/blob/master/README.md#refresh-token)
+
+### Changed
+
+- Added validation in *Create Signing Link* method: a document must have at least one role that does not have a fixed e-mail to create an invite link.
 
 ## [v1.4.0] - 2019-08-08