+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Added validation in *Create Signing Link* method: a document must have at least one role that does not have a fixed e-mail to create an invite link
+- Added optional ability to remove original documents after merging documents in *Merge Documents* feature
+- Updated example of merging documents in [README](https://github.com/signnow/SignNowNodeSDK/blob/master/README.md#merge-documents)
 
 ## [v1.4.0] - 2019-08-08